+++ tests/test_components.py
@@ -41,6 +41,12 @@ import pytest
 
 import yuyo
 
+try:
+    import tanjun
+
+except ModuleNotFoundError:
+    tanjun = None
+
 
 class TestComponentClient:
     def test___init___when_event_managed(self):
@@ -112,6 +118,49 @@ class TestComponentClient:
         mock_init.assert_called_once_with(server=mock_bot.interaction_server)
         mock_bot.add_shutdown_callback.assert_called_once_with(stub_client._on_stopping)
         mock_bot.add_startup_callback.assert_called_once_with(stub_client._on_starting)
+
+    @pytest.mark.skipif(tanjun is None, reason="Tanjun specific test")
+    def test_from_tanjun(self):
+        assert tanjun
+
+        mock_bot = mock.Mock()
+        mock_init = mock.Mock(return_value=None)
+
+        class StubClient(yuyo.ComponentClient):
+            __init__ = mock_init
+
+        stub_client = StubClient.from_tanjun(mock_bot)
+
+        assert isinstance(stub_client, StubClient)
+        mock_init.assert_called_once_with(event_manager=mock_bot.events, server=mock_bot.server)
+        mock_bot.set_type_dependency.assert_called_once_with(
+            yuyo.ComponentClient, stub_client
+        )
+        mock_bot.add_client_callback.assert_not_called()
+
+    @pytest.mark.skipif(tanjun is None, reason="Tanjun specific test")
+    def test_from_tanjun_when_tanjun_managed(self):
+        assert tanjun
+
+        mock_bot = mock.Mock()
+        mock_init = mock.Mock(return_value=None)
+
+        class StubClient(yuyo.ComponentClient):
+            __init__ = mock_init
+
+        stub_client = StubClient.from_tanjun(mock_bot, tanjun_managed=True)
+
+        assert isinstance(stub_client, StubClient)
+        mock_init.assert_called_once_with(event_manager=mock_bot.events, server=mock_bot.server)
+        mock_bot.set_type_dependency.assert_called_once_with(
+            yuyo.ComponentClient, stub_client
+        )
+        mock_bot.add_client_callback.assert_has_called(
+            [
+                mock.call(tanjun.ClientCallbackNames.STARTING, stub_client.open),
+                mock.call(tanjun.ClientCallbackNames.CLOSING, stub_client.close)
+            ]
+        )
 
     @pytest.mark.asyncio()
     async def test__on_starting(self):