+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `from_tanjun` convenience classmethods for initalising from a Tanjun client to
+  [ReactionClient][yuyo.reactions.ReactionClient], [ModalClient][yuyo.modals.ModalClient],
+  [ServiceManager][yuyo.list_status.ServiceManager] and [ComponentClient][yuyo.modals.ModalClient].
+
 ### Fixed
 - Prefix matched custom IDs are now correctly lower priority for modals.